Updating the calculation for the carousel speed
This ensures that the carousel runs at a constant speed, regardless of
the number of items in the carousel. The speed is calculated based on the
total width of the items in the carousel and the desired speed in pixels
per second.

Added a padding parameter, which allows the user to specify the amount of space
to be added to the animation in order to ensure that the items are not cut off
at the edges of the carousel.